Rope move processor to new file
