+++ example_workflow/.slay_gen/processor_SplitText/user_dependencies.py
@@ -1,0 +1,3 @@
+import slay
+
+IMAGE_NUMPY = slay.Image().pip_install("numpy")
