+++ pyproject.toml
@@ -95,6 +95,7 @@ flask = "^2.3.3"
 types-requests = "2.31.0.2"
 libcst = "^1.2.0"
 redbaron = "^0.9.2"
+rope = "^1.12.0"
 
 [build-system]
 build-backend = "poetry.core.masonry.api"
