+++ slay/code_gen.py
@@ -1,5 +1,7 @@
 import logging
+import os
 import pathlib
+import shutil
 import textwrap
 from typing import Iterable
 
@@ -7,6 +9,9 @@ import black
 import isort
 import libcst
 import pydantic
+from rope.base import project as rope_project
+from rope.refactor import move as rope_move
+from rope.refactor import occurrences as rope_occurrences
 from slay import definitions, utils
 
 INDENT = " " * 4
@@ -231,6 +236,47 @@ def rewrite_processor_inits(
     return modified_tree
 
 
+def _rope_find_def_offset(project, source_module, symbol_name):
+    finder = rope_occurrences.Finder(project, symbol_name)
+    offset = -1
+    occurrences = finder.find_occurrences(source_module)
+    for occurrence in occurrences:
+        if occurrence.is_defined():
+            if offset >= 0:
+                raise ValueError("Multiple found")
+            offset = occurrence.offset
+    if offset < 0:
+        raise ValueError("Not found")
+    return offset
+
+
+def move_class_to_new_file(
+    project_root,
+    source_file_path,
+    target_file_path,
+    class_name,
+):
+    # TODO: all of this is totally unclear.
+    project = rope_project.Project(project_root, ropefolder=None)
+    source_module = project.get_resource(source_file_path)
+
+    offset = _rope_find_def_offset(project, source_module, class_name)
+
+    target_file = project.get_file(target_file_path)
+    if target_file.exists():
+        target_file.remove()
+    target_file.create()
+
+    target_resource = project.get_resource(target_file_path)
+
+    print(f"Moving `{class_name}` to `{target_file_path}`.")
+    move_refactoring = rope_move.create_move(project, source_module, offset)
+
+    changes = move_refactoring.get_changes(dest=target_resource)
+    # TODO: move other needed stuff from source module.
+    project.do(changes)
+
+
 ########################################################################################
 
 
@@ -238,15 +284,24 @@ def modify_source_file(
     file_path,
     processor_desrciptor: definitions.ProcessorAPIDescriptor,
 ):
+    wdir = os.path.dirname(file_path)
+
     with open(file_path, "r", encoding="utf-8") as source_file:
         source_code = source_file.read()
 
     source_tree = libcst.parse_module(source_code)
     source_tree = rewrite_processor_inits(source_tree, processor_desrciptor)
-
     modified_source_code = source_tree.code
 
     with open(file_path, "w", encoding="utf-8") as modified_file:
         modified_file.write(modified_source_code)
+
+    target_file_path = os.path.join(wdir, "model.py")
+    move_class_to_new_file(
+        wdir,
+        os.path.basename(file_path),
+        os.path.basename(target_file_path),
+        processor_desrciptor.processor_cls.__name__,
+    )
 
     _format_python_file(file_path)
