+++ slay/framework.py
@@ -350,30 +350,32 @@ class StubDescriptor(pydantic.BaseModel):
     url: str
 
 
-def _create_processor_dir(processor_descriptor):
-    workflow_filepath = os.path.abspath(sys.argv[0])
-    workflow_dir = os.path.dirname(workflow_filepath)
+def _create_processor_dir(workflow_root, processor_descriptor):
     processor_name = processor_descriptor.processor_cls.__name__
     processor_dir = os.path.join(
-        workflow_dir, ".slay_gen", f"processor_{processor_name}"
+        workflow_root, ".slay_gen", f"processor_{processor_name}"
     )
     os.makedirs(processor_dir, exist_ok=True)
     return processor_dir
 
 
 def create_remote_service(
+    workflow_root,
     processor_descriptor: definitions.ProcessorAPIDescriptor,
-    # stubs: list[StubDescriptor],
 ):
-    processor_dir = _create_processor_dir(processor_descriptor)
-    model_filepath = shutil.copy(
-        processor_descriptor.src_file, os.path.join(processor_dir, "model.py")
+    processor_dir = _create_processor_dir(workflow_root, processor_descriptor)
+    # TODO: copy other local deps.
+    processor_filepath = shutil.copy(
+        processor_descriptor.src_file,
+        os.path.join(processor_dir, "user_dependencies.py"),
     )
-    code_gen.modify_source_file(model_filepath, processor_descriptor)
-    return model_filepath
+    code_gen.modify_source_file(processor_filepath, processor_descriptor)
 
 
 def deploy_remotely(processors: Iterable[Type[definitions.ABCProcessor]]) -> None:
+    workflow_filepath = os.path.abspath(sys.argv[0])
+    workflow_root = os.path.dirname(workflow_filepath)
+
     needed_processors: set[definitions.ProcessorAPIDescriptor] = set()
 
     def add_needed_procssors(proc: definitions.ProcessorAPIDescriptor):
@@ -393,9 +395,9 @@ def deploy_remotely(processors: Iterable[Type[definitions.ABCProcessor]]) -> Non
     ]
 
     for processor_descriptor in ordered_processors:
-        processor_dir = _create_processor_dir(processor_descriptor)
+        processor_dir = _create_processor_dir(workflow_root, processor_descriptor)
         code_gen.generate_stubs_for_deps(
             os.path.join(processor_dir, "dependencies.py"),
             _global_processor_registry.get_dependencies(processor_descriptor),
         )
-        create_remote_service(processor_descriptor)
+        create_remote_service(workflow_root, processor_descriptor)
